requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
boto3==1.18.65
botocore==1.21.65
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==36.0.0
daphne==3.0.2
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
django-storages==1.12.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djongo==1.3.6
dnspython==2.1.0
drf-yasg==1.20.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupSafe==2.0.1
packaging==21.0
Pillow==8.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.1.0
pymongo==3.11.4
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.1
razorpay==1.2.0
requests==2.26.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
service-identity==21.1.0
sib-api-v3-sdk==7.2.0
six==1.16.0
sqlparse==0.2.4
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uritemplate==3.0.1
urllib3==1.26.6
zope.interface==5.4.0
